renames tags_access_rights
