+++ packages/postgres-database/src/simcore_postgres_database/models/services_tags.py
@@ -5,8 +5,7 @@ from .tags import tags
 
 services_tags = sa.Table(
     #
-    # - Tags assigned to a service
-    # - Many-to-many association
+    # Tags assigned to a service (many-to-many relation)
     #
     "services_tags",
     metadata,
