+++ packages/postgres-database/tests/test_utils_tags.py
@@ -11,7 +11,7 @@ import sqlalchemy as sa
 from aiopg.sa.connection import SAConnection
 from aiopg.sa.result import RowProxy
 from pytest_simcore.helpers.postgres_tags import create_tag, create_tag_access
-from simcore_postgres_database.models.tags_to_groups import tags_to_groups
+from simcore_postgres_database.models.tags_access_rights import tags_to_groups
 from simcore_postgres_database.models.users import UserRole, UserStatus
 from simcore_postgres_database.tags_sql import (
     add_tag_to_project_stmt,
